+++ sdk/src/wallet/account/operations/participation/mod.rs
@@ -60,10 +60,11 @@ impl Account {
         // TODO: Could use the address endpoint in the future when https://github.com/iotaledger/inx-participation/issues/50 is done.
 
         let mut spent_cached_outputs = self
+            .wallet
             .storage_manager
             .lock()
             .await
-            .get_cached_participation_output_status(self.read().await.index)
+            .get_cached_participation_output_status(self.details().await.index)
             .await?;
         let restored_spent_cached_outputs_len = spent_cached_outputs.len();
         log::debug!(
@@ -208,10 +209,11 @@ impl Account {
         );
         // Only store updated data if new outputs got added
         if spent_cached_outputs.len() > restored_spent_cached_outputs_len {
-            self.storage_manager
+            self.wallet
+                .storage_manager
                 .lock()
                 .await
-                .set_cached_participation_output_status(self.read().await.index, spent_cached_outputs)
+                .set_cached_participation_output_status(self.details().await.index, spent_cached_outputs)
                 .await?;
         }
 
@@ -236,8 +238,9 @@ impl Account {
     /// If event isn't found, the client from the account will be returned.
     pub(crate) async fn get_client_for_event(&self, id: &ParticipationEventId) -> crate::wallet::Result<Client> {
         log::debug!("[get_client_for_event]");
-        let account_index = self.read().await.index;
+        let account_index = self.details().await.index;
         let events = self
+            .wallet
             .storage_manager
             .lock()
             .await
@@ -265,8 +268,9 @@ impl Account {
         log::debug!("[remove_ended_participation_events]");
         let latest_milestone_index = self.client().get_info().await?.node_info.status.latest_milestone.index;
 
-        let account_index = self.read().await.index;
+        let account_index = self.details().await.index;
         let events = self
+            .wallet
             .storage_manager
             .lock()
             .await