+++ sdk/src/wallet/account/operations/transaction/high_level/burning_melting/destroy_alias.rs
@@ -80,7 +80,7 @@ impl Account {
         &self,
         alias_id: AliasId,
     ) -> crate::wallet::Result<(OutputId, Output)> {
-        let account_details = self.read().await;
+        let account_details = self.details().await;
         let token_supply = self.client.get_token_supply().await?;
 
         let (output_id, output_data) = account_details