+++ sdk/src/wallet/account/operations/transaction/mod.rs
@@ -179,7 +179,7 @@ impl Account {
             inputs,
         };
 
-        let mut account_details = self.write().await;
+        let mut account_details = self.details_mut().await;
 
         account_details.transactions.insert(transaction_id, transaction.clone());
         account_details.pending_transactions.insert(transaction_id);
@@ -194,7 +194,7 @@ impl Account {
 
     // unlock outputs
     async fn unlock_inputs(&self, inputs: &[InputSigningData]) -> crate::wallet::Result<()> {
-        let mut account_details = self.write().await;
+        let mut account_details = self.details_mut().await;
         for input_signing_data in inputs {
             let output_id = input_signing_data.output_id();
             account_details.locked_outputs.remove(output_id);